Avoid truncation of __fsword_t values
Commit UUID: ba012784-94ed-4481-be49-d42169b26003

+++ trunk/src/cp_mem.c
@@ -54,7 +54,7 @@ struct dest {
 };
 
 #if !defined(HAVE___FSWORD_T)
-typedef unsigned __fsword_t;
+typedef uint64_t __fsword_t;
 
 #endif
 static char **srcs;
